pyproject.toml
[project]
name = "impure-package"
version = "0.1.0"
description = "Add your description here"
authors = [ ]
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "deepseek>=1.0.0",
    "python-dotenv>=1.2.1",
]
